requirements.txt
numpy~=1.20.1
pandas~=1.2.4
psycopg2~=2.8.6
requests~=2.28.1
apache-airflow~=2.5.0
matplotlib~=3.3.4